rename to checkout sheet kit
