+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 - Implemented onRenderProcessGone to prevent app crashes when the render process is killed to reclaim memory
 - JavaDoc improvements
-- Add function to retrieve the currently applied ShopifyCheckoutKit configuration
+- Add function to retrieve the currently applied ShopifyCheckoutSheet configuration
 
 Note: breaking changes may be made during Dev Preview with minor version increments, but will receive a major version increments
 after release for General Availability.
