--- lib/src/main/java/com/shopify/checkoutkit/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Package containing all of the classes that comprise Shopify Checkout Kit.
- */
-package com.shopify.checkoutkit;
