+++ lib/src/main/java/com/shopify/checkoutsheetkit/CheckoutWebViewContainer.kt
@@ -1,18 +1,18 @@
 /*
  * MIT License
- *
+ * 
  * Copyright 2023-present, Shopify Inc.
- *
+ * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * 
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
- *
+ * 
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,13 +20,25 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkoutkit
+package com.shopify.checkoutsheetkit
 
-fun withPreloadingEnabled(block: () -> Unit) {
-    try {
-        ShopifyCheckoutKit.configure { it.preloading = Preloading(enabled = true) }
-        block()
-    } finally {
-        ShopifyCheckoutKit.configure { it.preloading = Preloading(enabled = false) }
+import android.content.Context
+import android.util.AttributeSet
+import android.view.View
+import android.widget.FrameLayout
+
+internal class CheckoutWebViewContainer @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0,
+    defStyleRes: Int = 0
+) : FrameLayout(context, attrs, defStyleAttr, defStyleRes) {
+
+    // Clear the cache whenever the WebView is removed from it's container
+    override fun onViewRemoved(child: View?) {
+        super.onViewRemoved(child)
+        if (child is CheckoutWebView) {
+            CheckoutWebView.clearCache()
+        }
     }
 }
