+++ lib/src/main/java/com/shopify/checkoutsheetkit/Configuration.kt
@@ -1,18 +1,18 @@
 /*
  * MIT License
- *
+ * 
  * Copyright 2023-present, Shopify Inc.
- *
+ * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * 
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
- *
+ * 
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,15 +20,25 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkoutkit
+package com.shopify.checkoutsheetkit
 
 /**
- * Interface used to update the SDK configuration.
+ * Configuration for Shopify Checkout Sheet Kit.
+ *
+ * Allows:
+ * - Enabling/disabling preloading,
+ * - Specifying the colorScheme that should be used for checkout.
  */
-public fun interface ConfigurationUpdater {
-    /**
-     * Configuration block that the SDK will run to update the SDK settings.
-     * @param configuration object that you can modify (via setters) to update the settings.
-     */
-    public fun configure(configuration: Configuration)
-}
+public data class Configuration internal constructor(
+    var colorScheme: ColorScheme = ColorScheme.Automatic(),
+    var preloading: Preloading = Preloading(),
+)
+
+/**
+ * Configuration related to preloading.
+ *
+ * Initially allows toggling the preloading feature.
+ */
+public data class Preloading(
+    val enabled: Boolean = false
+)
