+++ lib/src/main/java/com/shopify/checkoutsheetkit/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package containing all of the classes that comprise Shopify Checkout Sheet Kit.
+ */
+package com.shopify.checkoutsheetkit;
