+++ lib/src/test/java/com/shopify/checkoutsheetkit/CheckoutDialogTest.kt
@@ -20,7 +20,7 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkoutkit
+package com.shopify.checkoutsheetkit
 
 import android.app.Dialog
 import android.graphics.drawable.ColorDrawable
@@ -53,7 +53,7 @@ class CheckoutDialogTest {
 
     @Test
     fun `shows dialog when present is called`() {
-        ShopifyCheckoutKit.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
 
@@ -62,7 +62,7 @@ class CheckoutDialogTest {
 
     @Test
     fun `when dialog is presented checkoutView is added to the container`() {
-        ShopifyCheckoutKit.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
 
@@ -75,7 +75,7 @@ class CheckoutDialogTest {
 
     @Test
     fun `cancel() removes checkoutView from the container`() {
-        ShopifyCheckoutKit.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         assertThat(containerChildCount(dialog)).isEqualTo(1)
@@ -92,7 +92,7 @@ class CheckoutDialogTest {
     fun `cancel() removes checkoutView from the container and invalidates the cache`() {
         withPreloadingEnabled {
             val url = "https://shopify.com"
-            ShopifyCheckoutKit.present(url, activity, NoopEventProcessor())
+            ShopifyCheckoutSheet.present(url, activity, NoopEventProcessor())
             val dialog = ShadowDialog.getLatestDialog()
             val preCancelView = CheckoutWebView.cacheableCheckoutView(url, activity)
 
@@ -106,7 +106,7 @@ class CheckoutDialogTest {
 
     @Test
     fun `clicking close invokes cancel(), removing checkoutView from the container`() {
-        ShopifyCheckoutKit.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         assertThat(containerChildCount(dialog)).isEqualTo(1)
@@ -124,9 +124,9 @@ class CheckoutDialogTest {
     @Test
     fun `sets header background color based on current configuration`() {
         val customColors = customColors()
-        ShopifyCheckoutKit.configuration.colorScheme = ColorScheme.Web(customColors)
+        ShopifyCheckoutSheet.configuration.colorScheme = ColorScheme.Web(customColors)
 
-        ShopifyCheckoutKit.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         val header = dialog.findViewById<Toolbar>(R.id.checkoutSdkHeader)
@@ -139,9 +139,9 @@ class CheckoutDialogTest {
     @Test
     fun `sets webview container background color based on current configuration`() {
         val customColors = customColors()
-        ShopifyCheckoutKit.configuration.colorScheme = ColorScheme.Web(customColors)
+        ShopifyCheckoutSheet.configuration.colorScheme = ColorScheme.Web(customColors)
 
-        ShopifyCheckoutKit.present("https://shopify.com", activity, NoopEventProcessor())
+        ShopifyCheckoutSheet.present("https://shopify.com", activity, NoopEventProcessor())
 
         val dialog = ShadowDialog.getLatestDialog()
         val webViewContainer = dialog.findViewById<FrameLayout>(R.id.checkoutSdkContainer)
