+++ lib/src/test/java/com/shopify/checkoutsheetkit/Helpers.kt
@@ -1,18 +1,18 @@
 /*
  * MIT License
- * 
+ *
  * Copyright 2023-present, Shopify Inc.
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in all
  * copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,23 +20,13 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.shopify.checkoutkit
+package com.shopify.checkoutsheetkit
 
-import android.util.Log
-
-/**
- * Wrap Log class static methods to allow testing
- */
-public class LogWrapper {
-    public fun w(tag: String, msg: String) {
-        Log.w(tag, msg)
-    }
-
-    public fun e(tag: String, msg: String) {
-        Log.e(tag, msg)
-    }
-
-    public fun e(tag: String, msg: String, throwable: Throwable) {
-        Log.e(tag, msg, throwable)
+fun withPreloadingEnabled(block: () -> Unit) {
+    try {
+        ShopifyCheckoutSheet.configure { it.preloading = Preloading(enabled = true) }
+        block()
+    } finally {
+        ShopifyCheckoutSheet.configure { it.preloading = Preloading(enabled = false) }
     }
 }
