+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/cart/CartViewModel.kt
@@ -28,8 +28,8 @@ import androidx.lifecycle.viewModelScope
 import com.shopify.buy3.Storefront
 import com.shopify.buy3.Storefront.Cart
 import com.shopify.buy3.Storefront.CartLineInput
-import com.shopify.checkoutkit.CheckoutEventProcessor
-import com.shopify.checkoutkit.ShopifyCheckoutKit
+import com.shopify.checkoutsheetkit.CheckoutEventProcessor
+import com.shopify.checkoutsheetkit.ShopifyCheckoutSheet
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.common.client.StorefrontClient
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.settings.PreferencesManager
 import com.shopify.graphql.support.ID
@@ -77,7 +77,7 @@ class CartViewModel(
         activity: ComponentActivity,
         eventProcessor: CheckoutEventProcessor
     ) {
-        ShopifyCheckoutKit.present(url, activity, eventProcessor)
+        ShopifyCheckoutSheet.present(url, activity, eventProcessor)
     }
 
     private fun performCartLinesAdd(cartID: ID, variant: ID, onComplete: OnComplete) {
