+++ samples/MobileBuyIntegration/app/src/main/java/com/shopify/checkout_sdk_mobile_buy_integration_sample/settings/ColorSchemeSection.kt
@@ -45,10 +45,10 @@ import androidx.compose.ui.text.intl.Locale
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.shopify.checkoutkit.ColorScheme
-import com.shopify.checkoutkit.Colors
+import com.shopify.checkoutsheetkit.ColorScheme
+import com.shopify.checkoutsheetkit.Colors
 import com.shopify.checkout_sdk_mobile_buy_integration_sample.R
-import com.shopify.checkoutkit.Color
+import com.shopify.checkoutsheetkit.Color
 
 @Composable
 fun ColorSchemeSection(
