+++ samples/README.md
@@ -12,9 +12,9 @@ to `.env` and update the contained values to match your Shopify storefront.
 
 This project demonstrates how to use
 the [Mobile Buy SDK](https://github.com/Shopify/mobile-buy-sdk-android) in conjunction with
-the `ShopifyCheckoutKit` library.
+the `Shopify Checkout Sheet Kit` library.
 
 ### SimpleAppIntegration
 
-This project demonstrates how to use the `ShopifyCheckoutKit` library with a custom built Storefront
+This project demonstrates how to use the `Shopify Checkout Sheet Kit` library with a custom built Storefront
 GraphQL API client.
