+++ settings.gradle
@@ -14,6 +14,6 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "checkout-kit"
+rootProject.name = "checkout-sheet-kit"
 
 include ':lib'
